vunit/requirements.txt
paho-mqtt~=1.6.1
pymongo~=4.1.1
pyserial~=3.5
smbus2~=0.4.1
RPi.GPIO
Adafruit_DHT
GreenPonik_BH1750